ddpg trainin

+++ pfrl/utils/concat_obs_and_goal.py
@@ -6,4 +6,4 @@ def concat_obs_and_goal(obs_dict):
     concatenates the observation
     and goal in the observation dictionary.
     """
-    return np.concatenate((obs_dict['observation'], obs_dict['desired_goal']))
+    return np.concatenate((obs_dict['observation'], obs_dict['desired_goal'])).astype(np.float32)
